test(gateway): dedup Lazer live-test bootstrap into a shared helper (ENG-398)
Both live-network Lazer tests read the same env vars and build the same
LazerSourceConfig. The fixture-capture is a unit test (needs crate-internal
lazer_wire items); the payload-fetch smoke test only used the public
LazerPayloadSource, so it had no reason to be a separate integration-test
binary. Move it into lazer_client::tests and share one in-module
live_config_from_env() helper instead of duplicating the bootstrap.
